Fixed type for window resolution

+++ src/main.rs
@@ -5,15 +5,15 @@ use bevy::prelude::*;
 use seldom_state::prelude::*;
 use entities::player::PlayerPlugin;
 
-const WINDOW_WIDTH: f32 = 1280.0;
-const WINDOW_HEIGHT: f32 = 720.0;
+const WINDOW_WIDTH: u32 = 1280;
+const WINDOW_HEIGHT: u32 = 720;
 const WINDOW_TITLE: &str = "Hello World";
 
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins.set(WindowPlugin {
             primary_window: Some(Window {
-                resolution: (WINDOW_WIDTH as u32, WINDOW_HEIGHT as u32).into(),
+                resolution: (WINDOW_WIDTH, WINDOW_HEIGHT).into(),
                 title: WINDOW_TITLE.into(),
                 ..default()
             }),
